+++ ReleaseNotes.md
@@ -1,12 +1,20 @@
 # ODC Release Notes
 
-## v0.4 (not yet released)
+## v0.6 (NOT YET RELEASED)
+
+### ODC common
+Added: new Run request which combines Initialize, Submit and Activate. Run request always creates a new DDS session.    
+
+
+
+## v0.4 (2020-07-01)
 
 ### ODC common
 Added: possibility to attach to the running DDS session.    
 Added: DDS session ID in each reply.    
 Added: set request timeout via command line interface.    
 Added: unit tests.    
+Modified: minimum required DDS version is 3.4    
 
 
 ## v0.2 (2020-05-13)